+++ CHANGELOG.md
@@ -2,10 +2,16 @@
 
 ## [Unreleased] - TODO
 
+### Changed
+
 - Update dependencies
-  - `Kotlin` to `1.8.22`.
+  - `Kotlin` to `1.9.0`.
   - `KotlinX Coroutines` to `1.7.2`.
   - `Gradle` to `8.2`.
+
+### Added
+
+- Add `Flow.repeat` operator.
 
 ## [0.6.1] - May 18, 2023